+++ CHANGELOG.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.0] - 2024-04-02
+## [0.2.1] - 2025-04-03
+
+### Added
+
+- The ability to select embedding functions when creating collections (default, cohere, openai, jina, voyageai, roboflow)
+
+### Changed
+- Upgraded to v1.0.0 of Chroma
+- Fix dotenv path support during argparse
+
+## [0.2.0] - 2025-04-02
 
 ### Added
 - New `delete_document` tool for removing documents from collections